requirements.txt
asgiref==3.11.0
Django==5.2.9
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.4
drf-nested-routers==0.95.0
requests==2.32.5
django-cors-headers==4.9.0